refactor keeper unit tests
